fix: close+reopen popover on navigation — only correct NSPopover anchor fix
NSPopover cannot reposition its window after show(). The only correct
approach: close, swap content+size while closed, reopen on next runloop
tick so the anchor calculation uses the correct status button position.
