+++ Sources/RunnerBar/AppDelegate.swift
@@ -5,35 +5,12 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
 
     private var statusItem: NSStatusItem?
     private var popover: NSPopover?
-    private var hc: NSHostingController<PopoverView>?
+    private var hc: NSHostingController<AnyView>?
     private let observable = RunnerStoreObservable()
 
-    // Navigation state lives here in AppDelegate so we can re-anchor
-    // the popover (via show(relativeTo:of:preferredEdge:)) whenever the
-    // content size changes. Toggling this calls showPopover() which
-    // sets the new contentSize BEFORE showing, preventing the (0,0) jump.
-    var selectedJob: ActiveJob? = nil {
-        didSet {
-            // Update the hosted root view with the new selection.
-            if let hc {
-                hc.rootView = PopoverView(
-                    store: observable,
-                    selectedJob: selectedJob,
-                    onSelectJob: { [weak self] job in self?.selectedJob = job },
-                    onBack: { [weak self] in self?.selectedJob = nil }
-                )
-            }
-            // Re-show the popover so it re-anchors at the correct position
-            // with the updated contentSize. This is the only reliable fix.
-            if popover?.isShown == true {
-                showPopover()
-            }
-        }
-    }
+    private static let popoverWidth: CGFloat = 320
 
     func applicationDidFinishLaunching(_ notification: Notification) {
-        log("AppDelegate \u203a applicationDidFinishLaunching")
-
         statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.squareLength)
         if let button = statusItem?.button {
             button.image = makeStatusIcon(for: .allOffline)
@@ -41,12 +18,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
             button.target = self
         }
 
-        let rootView = PopoverView(
-            store: observable,
-            selectedJob: nil,
-            onSelectJob: { [weak self] job in self?.selectedJob = job },
-            onBack: { [weak self] in self?.selectedJob = nil }
-        )
+        let rootView = AnyView(mainView())
         let hc = NSHostingController(rootView: rootView)
         hc.sizingOptions = []
         self.hc = hc
@@ -54,24 +26,55 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         let popover = NSPopover()
         popover.behavior              = .transient
         popover.animates              = false
-        popover.contentSize           = Self.mainSize
         popover.contentViewController = hc
         self.popover = popover
 
         RunnerStore.shared.onChange = { [weak self] in
             guard let self else { return }
-            log("AppDelegate \u203a onChange")
             self.statusItem?.button?.image = makeStatusIcon(for: RunnerStore.shared.aggregateStatus)
             self.observable.reload()
         }
-
         RunnerStore.shared.start()
     }
 
-    // MARK: - Sizes
+    // MARK: - View factories
 
-    private static let mainSize   = NSSize(width: 320, height: 420)
-    private static let detailSize = NSSize(width: 320, height: 420)
+    private func mainView() -> some View {
+        PopoverMainView(store: observable, onSelectJob: { [weak self] job in
+            self?.navigateTo(job: job)
+        })
+    }
+
+    private func detailView(job: ActiveJob) -> some View {
+        JobDetailView(job: job, onBack: { [weak self] in
+            self?.navigateBack()
+        })
+    }
+
+    // MARK: - Navigation
+
+    private func navigateTo(job: ActiveJob) {
+        // 1. Close while open so next show() re-anchors correctly.
+        popover?.performClose(nil)
+        // 2. Swap content and size while closed.
+        let view = AnyView(detailView(job: job))
+        hc?.rootView = view
+        let size = NSSize(width: Self.popoverWidth, height: 460)
+        popover?.contentSize = size
+        hc?.view.setFrameSize(size)
+        // 3. Re-open on next runloop tick.
+        DispatchQueue.main.async { [weak self] in self?.showPopover() }
+    }
+
+    private func navigateBack() {
+        popover?.performClose(nil)
+        let view = AnyView(mainView())
+        hc?.rootView = view
+        let size = NSSize(width: Self.popoverWidth, height: 420)
+        popover?.contentSize = size
+        hc?.view.setFrameSize(size)
+        DispatchQueue.main.async { [weak self] in self?.showPopover() }
+    }
 
     // MARK: - Toggle
 
@@ -80,6 +83,10 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         if popover.isShown {
             popover.performClose(nil)
         } else {
+            // Ensure main view is loaded with correct size.
+            let size = NSSize(width: Self.popoverWidth, height: 420)
+            popover.contentSize = size
+            hc?.view.setFrameSize(size)
             showPopover()
         }
     }
@@ -87,15 +94,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     private func showPopover() {
         guard let button = statusItem?.button,
               button.window != nil,
-              let popover, let hc else { return }
-
-        // Set size BEFORE showing so the anchor calculation uses correct size.
-        let size = (selectedJob == nil) ? Self.mainSize : Self.detailSize
-        popover.contentSize = size
-        hc.view.setFrameSize(size)
-        hc.view.layoutSubtreeIfNeeded()
-
-        log("AppDelegate \u203a showPopover size=\(size) isShown=\(popover.isShown)")
+              let popover else { return }
         popover.show(relativeTo: button.bounds, of: button, preferredEdge: .maxY)
         popover.contentViewController?.view.window?.makeKey()
     }
