+++ Sources/RunnerBar/PopoverMainView.swift
@@ -1,0 +1,226 @@
+import SwiftUI
+import ServiceManagement
+
+/// The main list screen of the popover.
+/// Navigation is driven by AppDelegate (close+reopen) so this view
+/// has no knowledge of JobDetailView at all.
+struct PopoverMainView: View {
+    @ObservedObject var store: RunnerStoreObservable
+    let onSelectJob: (ActiveJob) -> Void
+
+    @State private var newScope = ""
+    @State private var launchAtLogin = LoginItem.isEnabled
+    @State private var isAuthenticated = (githubToken() != nil)
+    @State private var tick = 0
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 0) {
+
+            // ── Header
+            HStack {
+                Text("RunnerBar v0.8")
+                    .font(.headline)
+                    .foregroundColor(.secondary)
+                Spacer()
+                if isAuthenticated {
+                    HStack(spacing: 4) {
+                        Circle().fill(Color.green).frame(width: 8, height: 8)
+                        Text("Authenticated").font(.caption).foregroundColor(.secondary)
+                    }
+                } else {
+                    Button(action: signInWithGitHub) {
+                        HStack(spacing: 4) {
+                            Circle().fill(Color.orange).frame(width: 8, height: 8)
+                            Text("Sign in with GitHub").font(.caption).foregroundColor(.orange)
+                        }
+                    }.buttonStyle(.plain)
+                }
+            }
+            .padding(.horizontal, 12)
+            .padding(.top, 12)
+            .padding(.bottom, 8)
+
+            Divider()
+
+            // ── Active Jobs
+            Text("Active Jobs")
+                .font(.caption).foregroundColor(.secondary)
+                .padding(.horizontal, 12).padding(.top, 8).padding(.bottom, 2)
+
+            if store.jobs.isEmpty {
+                Text("No active jobs")
+                    .font(.caption).foregroundColor(.secondary)
+                    .padding(.horizontal, 12).padding(.vertical, 4).padding(.bottom, 2)
+            } else {
+                ForEach(store.jobs.prefix(3)) { job in
+                    Button(action: { onSelectJob(job) }) {
+                        HStack(spacing: 8) {
+                            jobDot(for: job)
+                            Text(job.name)
+                                .font(.system(size: 12))
+                                .foregroundColor(job.isDimmed ? .secondary : .primary)
+                                .lineLimit(1).truncationMode(.tail)
+                            Spacer()
+                            if job.isDimmed {
+                                Text(conclusionLabel(for: job))
+                                    .font(.caption).foregroundColor(conclusionColor(for: job))
+                                    .frame(width: 76, alignment: .trailing)
+                            } else {
+                                Text(jobStatusLabel(for: job))
+                                    .font(.caption).foregroundColor(jobStatusColor(for: job))
+                                    .frame(width: 76, alignment: .trailing)
+                            }
+                            Text(job.isDimmed ? job.elapsed : elapsedLive(for: job))
+                                .font(.caption.monospacedDigit()).foregroundColor(.secondary)
+                                .frame(width: 40, alignment: .trailing)
+                            Image(systemName: "chevron.right")
+                                .font(.caption2).foregroundColor(.secondary)
+                        }
+                        .padding(.horizontal, 12).padding(.vertical, 3)
+                    }
+                    .buttonStyle(.plain)
+                }
+                .padding(.bottom, 6)
+            }
+
+            Divider()
+
+            // ── Local runners
+            if !store.runners.isEmpty {
+                Text("Local runners")
+                    .font(.caption).foregroundColor(.secondary)
+                    .padding(.horizontal, 12).padding(.top, 8).padding(.bottom, 2)
+                ForEach(store.runners, id: \.id) { runner in
+                    HStack(spacing: 8) {
+                        Circle().fill(dotColor(for: runner)).frame(width: 8, height: 8)
+                        Text(runner.name).font(.system(size: 13)).lineLimit(1)
+                        Spacer()
+                        Text(runner.displayStatus)
+                            .font(.caption).foregroundColor(.secondary).lineLimit(1).fixedSize()
+                    }
+                    .padding(.horizontal, 12).padding(.vertical, 5)
+                }
+                Divider()
+            }
+
+            // ── Scopes
+            VStack(alignment: .leading, spacing: 4) {
+                Text("Scopes")
+                    .font(.caption).foregroundColor(.secondary)
+                    .padding(.horizontal, 12).padding(.top, 8)
+                ForEach(ScopeStore.shared.scopes, id: \.self) { scope in
+                    HStack {
+                        Text(scope).font(.system(size: 12))
+                        Spacer()
+                        Button(action: {
+                            ScopeStore.shared.remove(scope)
+                            store.reload()
+                        }) {
+                            Image(systemName: "minus.circle").foregroundColor(.red)
+                        }.buttonStyle(.plain)
+                    }
+                    .padding(.horizontal, 12).padding(.vertical, 2)
+                }
+                HStack {
+                    TextField("owner/repo or org", text: $newScope)
+                        .textFieldStyle(.roundedBorder).font(.system(size: 12))
+                        .onSubmit { submitScope() }
+                    Button(action: submitScope) {
+                        Image(systemName: "plus.circle")
+                    }.buttonStyle(.plain)
+                     .disabled(newScope.trimmingCharacters(in: .whitespaces).isEmpty)
+                }
+                .padding(.horizontal, 12).padding(.vertical, 4)
+            }
+
+            Divider()
+
+            Toggle(isOn: $launchAtLogin) {
+                Text("Launch at login").font(.system(size: 13))
+            }
+            .toggleStyle(.checkbox)
+            .padding(.horizontal, 12).padding(.vertical, 8)
+            .onChange(of: launchAtLogin) { _ in LoginItem.toggle() }
+
+            Divider()
+
+            Button(action: { NSApplication.shared.terminate(nil) }) {
+                HStack {
+                    Image(systemName: "xmark.square")
+                    Text("Quit")
+                }.font(.system(size: 13))
+            }
+            .buttonStyle(.plain)
+            .keyboardShortcut("q", modifiers: .command)
+            .padding(.horizontal, 12).padding(.vertical, 8)
+
+            Spacer(minLength: 0)
+        }
+        .onReceive(store.objectWillChange) { isAuthenticated = (githubToken() != nil) }
+        .onAppear {
+            Timer.scheduledTimer(withTimeInterval: 1, repeats: true) { _ in tick += 1 }
+        }
+    }
+
+    private func elapsedLive(for job: ActiveJob) -> String { job.elapsed }
+
+    @ViewBuilder
+    private func jobDot(for job: ActiveJob) -> some View {
+        Circle()
+            .fill(job.isDimmed ? Color.secondary : jobDotColor(for: job))
+            .frame(width: 7, height: 7)
+    }
+    private func jobDotColor(for job: ActiveJob) -> Color {
+        switch job.status {
+        case "in_progress": return .yellow
+        case "queued":      return .gray
+        default:            return .secondary
+        }
+    }
+    private func jobStatusLabel(for job: ActiveJob) -> String {
+        switch job.status {
+        case "in_progress": return "In Progress"
+        case "queued":      return "Queued"
+        default:            return "Done"
+        }
+    }
+    private func jobStatusColor(for job: ActiveJob) -> Color {
+        switch job.status {
+        case "in_progress": return .yellow
+        default:            return .secondary
+        }
+    }
+    private func conclusionLabel(for job: ActiveJob) -> String {
+        switch job.conclusion {
+        case "success":   return "\u2713 success"
+        case "failure":   return "\u2717 failure"
+        case "cancelled": return "\u2296 cancelled"
+        case "skipped":   return "\u2212 skipped"
+        default:          return job.conclusion ?? "done"
+        }
+    }
+    private func conclusionColor(for job: ActiveJob) -> Color {
+        switch job.conclusion {
+        case "success": return .green
+        case "failure": return .red
+        default:        return .secondary
+        }
+    }
+    private func dotColor(for runner: Runner) -> Color {
+        if runner.status != "online" { return .gray }
+        return runner.busy ? .yellow : .green
+    }
+    private func signInWithGitHub() {
+        let script = "tell application \"Terminal\" to do script \"gh auth login\""
+        NSAppleScript(source: script)?.executeAndReturnError(nil)
+        NSWorkspace.shared.open(URL(fileURLWithPath: "/System/Applications/Utilities/Terminal.app"))
+    }
+    private func submitScope() {
+        let trimmed = newScope.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !trimmed.isEmpty else { return }
+        ScopeStore.shared.add(trimmed)
+        RunnerStore.shared.start()
+        store.reload()
+        newScope = ""
+    }
+}
